fix: added some cases for yields and changed floating point types

+++ src/threads/fixed-point.h
@@ -6,17 +6,17 @@
 #define F 16384
 
 static inline
-int int_to_fp (int n){
+int64_t int_to_fp (int n){
     return n*F;
 }
 
 static inline
-int fp_to_int_truncated (int x) {
+int fp_to_int_truncated (int64_t x) {
     return x / F;
 }
 
 static inline
-int fp_to_int_rounded (int x){
+int fp_to_int_rounded (int64_t x){
     int offset;
     if (x < 0) offset = -(F/2);
     else offset = F/2;
@@ -24,28 +24,28 @@ int fp_to_int_rounded (int x){
 }
 
 static inline
-int add_n_to_fp (int n, int x) {
+int64_t add_n_to_fp (int n, int64_t x) {
     return x + n * F;
 }
 
 static inline
-int subtract_n_from_fp(int n, int x) {
+int64_t subtract_n_from_fp(int n, int64_t x) {
     return x - n * F;
 }
 
 static inline
-int subtract_fp_from_n(int n, int x) {
+int64_t subtract_fp_from_n(int n, int64_t x) {
     return n * F - x;
 }
 
 static inline
-int multiply_fp(int x, int y){
-    return (int) (((int64_t) x) * y /F);
+int64_t multiply_fp(int64_t x, int64_t y){
+    return ((int64_t) x) * y / F;
 }
 
 static inline
-int divide_fp(int x, int y){
-    return (int) (((int64_t) x) * F / y);
+int64_t divide_fp(int64_t x, int64_t y){
+    return ((int64_t) x) * F / y;
 }
 
 #endif /* threads/fixed-point.h */
